refactor: move TestBalances to a helpers folder in src
